--- src/Condition.java
@@ -1,5 +1,0 @@
-
-public abstract class Condition {
-    public abstract boolean checksCondition(Departement d);
-    protected abstract void setAttributes(Departement d);
-}
