--- src/main/java/com/example/geodevineur/Cardinal.java
@@ -1,5 +1,0 @@
-package com.example.geodevineur;
-
-public enum Cardinal {
-    W, NW, N, NE, E, SE, S, SW, C;
-}
