--- src/main/java/com/example/geodevineur/Politic.java
@@ -1,5 +1,0 @@
-package com.example.geodevineur;
-
-public enum Politic {
-    GAUCHE, DROITE, CENTRE;
-}
