+++ src/main/java/com/example/geodevineur/departement/Cardinal.java
@@ -1,3 +1,5 @@
+package com.example.geodevineur.departement;
+
 public enum Cardinal {
     W, NW, N, NE, E, SE, S, SW, C;
 }
