+++ src/main/java/com/example/geodevineur/departement/Politic.java
@@ -1,3 +1,5 @@
+package com.example.geodevineur.departement;
+
 public enum Politic {
     GAUCHE, DROITE, CENTRE;
 }
