+++ src/main/java/com/example/geodevineur/departement/condition/CardinalCond.java
@@ -1,3 +1,8 @@
+package com.example.geodevineur.departement.condition;
+
+import com.example.geodevineur.departement.Cardinal;
+import com.example.geodevineur.departement.Departement;
+
 public class CardinalCond extends Condition{
     private Cardinal cardinal;
