+++ src/main/java/com/example/geodevineur/departement/condition/Condition.java
@@ -1,4 +1,7 @@
-package com.example.geodevineur;
+package com.example.geodevineur.departement.condition;
+
+import com.example.geodevineur.departement.Departement;
+
 
 public abstract class Condition {
     public abstract boolean checksCondition(Departement d);
