+++ src/main/java/com/example/geodevineur/departement/condition/PopulationCond.java
@@ -1,4 +1,8 @@
+package com.example.geodevineur.departement.condition;
+
 import java.util.Random;
+import com.example.geodevineur.departement.Departement;
+
 
 public class PopulationCond extends Condition {
     private Random random;
