+++ src/main/java/com/example/geodevineur/departement/condition/SurfaceCond.java
@@ -1,14 +1,16 @@
-package com.example.geodevineur;
+package com.example.geodevineur.departement.condition;
+
+import com.example.geodevineur.departement.Departement;
 
 import java.util.Random;
 
-public class PopulationCond extends Condition {
+public class SurfaceCond extends Condition {
     private Random random;
     private Departement compareDep;
     private boolean isLess; // If more, isLess is false
     private int threshold;
     
-    public PopulationCond (Departement dep_, Departement compareDep_) {
+    public SurfaceCond (Departement dep_, Departement compareDep_) {
         random = new Random();
         setAttributes(dep_);
         double probaCompare = .3;
@@ -22,22 +24,27 @@ public class PopulationCond extends Condition {
 
     protected void setAttributes(Departement dep) {
         if (compareDep != null) {
-            threshold = compareDep.getPopulation();
-            if (threshold < dep.getPopulation()) {
-                isLess = false;
-            } else {
-                isLess = true;
-            }
+            threshold = compareDep.getSurface();
         } else {
-            // create appropriate thresholds
+            int rand = random.nextInt(2);
+            if (rand == 0) {
+                threshold = 5500;
+            } else {
+                threshold = 6500;
+            }
+        }
+        if (threshold < dep.getSurface()) {
+            isLess = false;
+        } else {
+            isLess = true;
         }
     }
 
     public boolean checksCondition(Departement dep) {
         if (isLess) {
-            return dep.getPopulation() < threshold;
+            return dep.getSurface() < threshold;
         } else {
-            return dep.getPopulation() > threshold;
+            return dep.getSurface() > threshold;
         }
     }
 }
